世帯員毎に共通する療育手帳等級の選択項目コンポーネントをIntellectualDisabilityとして切り出した
